2-openai/requirements.txt
langchain
python-dotenv
langchain-openai
langchain_community
streamlit
